feat:메인페이지 코디상품 불러오기

+++ src/main/java/com/example/lookkit/product/ProductVO.java
@@ -17,7 +17,7 @@ public class ProductVO {
     private Long codiId;
     private String productName;
     private String productDescription;
-    private double productPrice;
+    private int productPrice;
     private int productStock;
     private String genderTarget;
     private String productThumbnail;
